Update README.md
Spiral Framework is a High-Performance PHP/Go Full-Stack framework and PSR compatible components. The Framework execution model is based on a hybrid runtime where some services (GRPC, Queue, WebSockets, etc) handled on Application Server - RoadRunner. The PHP code of your application will stay in memory between the requests (the framework provides a set of instruments to avoid memory leaks). 

The spiral framework is the result of a decade of trial and error and currently used on production on multiple enterprise applications.